backend2/requirements.txt
Flask~=3.1.2
config~=0.5.1
python-dotenv~=1.2.1
uuid~=1.30
Werkzeug~=3.1.4
joblib~=1.5.3
pandas~=2.2.3
scikit-learn~=1.8.0
xgboost~=3.1.2
lightgbm~=4.6.0
alembic~=1.17.2
SQLAlchemy~=2.0.45